Add a scripts folder
